backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
passlib
PyJWT
email-validator
python-dotenv
bcrypt
mercadopago
websockets
tqdm
httpx